use staticcheck GH action
Don't run the staticcheck from the scripts because it needs to be
conditional upon the Go version being used. That is, in order to use newer
versions you need a newer Go.

Add a linter comment to the generated code to ignore usages of syscall.Syscall.
The Go team decided to no update existing code to use SyscallN so we will follow suit.